requirements.txt
pyxdg
coloredlogs
pandas<=1.5
openpyxl
requests
bibtexparser
python-docx
